Move IRenderTarget to Rendering
